docs: Split technical details and usage into separate files from README
Previously, all the information was included in the README.md file.
Now, the technical details are consolidated in the README.md,
while the usage details are documented separately in the docs.
The fpga-implementation.md file contains detailed instructions on how
to send and receive data through the FIFO in the FPGA.